[ADD] Add option to disable stock warning. [REM] Move sale_stcok_availability functionality to sale_stock_ux.
