+++ sale_stock_availability/README.rst
@@ -15,13 +15,6 @@ Stock availability in sales order line
 ======================================
 
 * Add two groups. One for seeing stock on sale orders and other to see only if or not available.
-* Add an option in warehouse to disable stock warning
-
-IMPORTANT:
-----------
-* This module could break some warnings as the ones implemented by 'warning' module
-* If you dont disable warning and give a user availbility to see only 'true/false' on sale order stock, he can see stock if the warning is raised
-
 
 Installation
 ============
@@ -44,7 +37,7 @@ Usage
 
 To use this module, you need to:
 
-#. If you want to add a product to the sales order line, the warning appears, if the stock of this product is less than the amount set.
+#. Just use the module.
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
