+++ sale_stock_availability/__manifest__.py
@@ -19,7 +19,7 @@
 ##############################################################################
 {
     'name': 'Stock availability in sales order line',
-    'version': '11.0.1.1.0',
+    'version': '11.0.1.2.0',
     'category': 'Tools',
     'author': 'Moldeo Interactive, ADHOC SA',
     'website': 'http://business.moldeo.coop http://adhoc.com.ar/',
@@ -32,7 +32,6 @@
     'data': [
         'security/sale_stock_availability_security.xml',
         'views/sale_order_views.xml',
-        'views/stock_warehouse_views.xml',
     ],
     'installable': True,
 }
