+++ sale_stock_availability/models/sale_order_line.py
@@ -22,11 +22,3 @@ class SaleOrderLine(models.Model):
             rec.virtual_available = rec.product_id.with_context(
                 warehouse=rec.order_id.warehouse_id.id
             ).virtual_available - rec.product_uom_qty
-
-    @api.onchange('product_uom_qty', 'product_uom', 'route_id')
-    def _onchange_product_id_check_availability(self):
-        res = super(SaleOrderLine,
-                    self)._onchange_product_id_check_availability()
-        if self.order_id.warehouse_id.disable_sale_stock_warning or False:
-            res.update({'warning': {}})
-        return res
