+++ sale_stock_ux/models/__init__.py
@@ -4,3 +4,4 @@
 ##############################################################################
 from . import sale_order
 from . import sale_order_line
+from . import stock_warehouse
