+++ sale_stock_ux/models/stock_warehouse.py
@@ -2,5 +2,12 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import stock_warehouse
-from . import sale_order_line
+from odoo import models, fields
+
+
+class StockWarehouse(models.Model):
+    _inherit = 'stock.warehouse'
+
+    disable_sale_stock_warning = fields.Boolean(
+        help='If true, the warning "Stock availability" is disable',
+    )
