+++ sale_stock_ux/views/stock_warehouse_views.xml
